### 2021-10-30 #3871 @whedon
[PRE REVIEW]: TriP: A Python package for the kinematic modeling of serial-parallel hybrid robots
**Submitting author:** @liquidcronos (<a href="http://orcid.org/0000-0002-7825-3476">Jan Baumgärtner</a>)
**Repository:** <a href="https://github.com/TriPed-Robot/TriP" target ="_blank">https://github.com/TriPed-Robot/TriP</a>
**Version:** v1.0.0
