### 2021-10-30 #3871 @danielskatz
👋 @liquidcronos - thanks for your submission.  Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.  You should hear back in a week or so.

In addition, you could work on the possibly missing DOIs that whedon suggests (and incorrect DOIs), but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references from branch paper` to check again, and the command `@whedon generate pdf from branch paper` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
