+++ paper.bib
@@ -1,23 +1,6 @@
 % Encoding: UTF-8
 
 
-@Book{robot_intro,
-  author    = {Bajd, Tadej and Mihelj, Matja{\v{z}} and Munih, Marko},
-  publisher = {Springer Science \& Business Media},
-  title     = {Introduction to robotics},
-  year      = {2013},
-}
-
-@InProceedings{ros,
-  author       = {Quigley, Morgan and Conley, Ken and Gerkey, Brian and Faust, Josh and Foote, Tully and Leibs, Jeremy and Wheeler, Rob and Ng, Andrew Y and others},
-  booktitle    = {ICRA workshop on open source software},
-  title        = {ROS: an open-source Robot Operating System},
-  year         = {2009},
-  number       = {3.2},
-  organization = {Kobe, Japan},
-  pages        = {5},
-  volume       = {3},
-}
 
 @Article{survey,
   author  = {Kumar, Shivesh and Woehrle, Hendrik and Fernández, José and Mueller, Andreas and Kirchner, Frank},
@@ -41,118 +24,6 @@
   publisher = {Springer},
 }
 
-@Article{hybrid_kinematics,
-  author         = {Chu, Anh My and Nguyen, Cong Dinh and Vu, Minh Hoan and Duong, Xuan Bien and Nguyen, Tien Anh and Le, Chi Hieu},
-  journal        = {Applied Sciences},
-  title          = {Kinematic and Dynamic Modelling for a Class of Hybrid Robots Composed of m Local Closed-Loop Linkages Appended to an n-Link Serial Manipulator},
-  year           = {2020},
-  issn           = {2076-3417},
-  number         = {7},
-  volume         = {10},
-  abstract       = {Recently, more and more hybrid robots have been designed to meet the increasing demand for a wide spectrum of applications. However, development of a general and systematic method for kinematic design and dynamic analysis for hybrid robots is rare. Most publications deal with the kinematic and dynamic issues for individual hybrid robots rather than any generalization. Hence, in this paper, we present a novel method for kinematic and dynamic modelling for a class of hybrid robots. First, a generic scheme for the kinematic design of a general hybrid robot mechanism is proposed. In this manner, the kinematic equation and the constraint equations for the robot class are derived in a generalized case. Second, in order to simplify the dynamic modelling and analysis of the complex hybrid robots, a Lemma about the analytical relationship among the generalized velocities of a hybrid robot system is proven in a generalized case as well. Last, examples of the kinematic and dynamic modelling of a newly designed hybrid robot are presented to demonstrate and validate the proposed method.},
-  article-number = {2567},
-  doi            = {10.3390/app10072567},
-  url            = {https://www.mdpi.com/2076-3417/10/7/2567},
-}
-
-@Book{modern_book,
-  author    = {Kevin M. Lynch and Frank C. Park},
-  publisher = {Cambridge University Press},
-  title     = {Modern Robotics: Mechanics, Planning, and Control},
-  year      = {2017},
-}
-
-@Misc{joints,
-  author  = {Er. Ashwini Kumar},
-  title   = {Theory of machines: Lesson 3},
-  year    = {2014},
-  journal = {http://ecoursesonline.iasri.res.in/mod/page/view.php?id=125546},
-}
-
-@Article{skrew,
-  author  = {Gallardo-Alvarado, Jaime},
-  journal = {Multibody System Dynamics},
-  title   = {Kinematics of a Hybrid Manipulator by Means of Screw Theory},
-  year    = {2005},
-  month   = {11},
-  pages   = {345-366},
-  volume  = {14},
-  doi     = {10.1007/s11044-005-4196-x},
-}
-
-@Article{88066,
-  author  = {Murray, J.J. and Lovell, G.H.},
-  journal = {IEEE Transactions on Robotics and Automation},
[…]
