### 2021-10-31 #3871 @liquidcronos
Hey @danielskatz thanks for considering TriP.
Most of these DOIs seem to be from bibtex entries that aren't used in the paper, I will will these Issues right away.
