### 2021-11-02 2ec8011
fixed spelling mistakes in docu
