### 2021-11-03 42b22fb
fixed symbolic transformation for endeffector description
