### 2021-11-03 86ef80d
removed unecessairy and missleading tips
added continuity measure to TriPed to prevent solution jumping in mappings
