+++ .github/workflows/python-publish.yml
@@ -39,3 +39,4 @@ jobs:
       with:
         user: __token__
         password: ${{ secrets.PYPI_API_TOKEN }}
+        packages_dir: src/dist/
