### 2021-11-12 #3871 @kyleniemeyer
Hi @liquidcronos, the editorial board has reviewed your submission and found that it meets our expectations for review, so this will continue. However, we do not have an editor available in this area right now, so I'm going to place this on our waitlist until someone becomes available (hopefully soon!)
