### 2021-11-16 3c0a30f
made triped joint names case compliant
