### 2021-11-16 1b2f629
fixed mapping arguments of triped model
