### 2021-11-23 #3871 @Kevin-Mattheus-Moerman
@liquidcronos I've checked today (I am AEiC on call this week) and all editors in the robotics domain are still fully occupied with other submissions. I'll check again in a couple of days and we will proceed to invite/assign an editor once they become available.
