### 2021-11-29 #3871 @danielskatz
👋 @sea-bass & @nnadeau - Would one or both of you be willing to review this JOSS submission?
