### 2021-11-29 #3871 @liquidcronos
> Thanks @liquidcronos - note that I've changed the title of this submission to fix a spelling error (seriell should be serial) - Can you do this in the paper.md as well?

Yeah, that the german spelling... Thanks for letting me know, it should be fixed now.
