### 2021-11-30 #3871 @sea-bass
> wave @sea-bass & @nnadeau - Would one or both of you be willing to review this JOSS submission?

Hi all -- unfortunately I am unable to review at the moment. Thank you for asking!
