### 2021-11-30 #3871 @danielskatz
👋 @adi3 and @trallard - Would one of you be willing to review this JOSS submission?
