### 2021-12-01 #3967 @whedon
[REVIEW]: TriP: A Python package for the kinematic modeling of serial-parallel hybrid robots
**Submitting author:** @liquidcronos (<a href="http://orcid.org/0000-0002-7825-3476">Jan Baumgärtner</a>)
**Repository:** https://github.com/TriPed-Robot/TriP
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.3
