### 2021-12-08 7ad98be
deleted unneeded keys() calls for iterators to improve speed
