+++ src/trip_kinematics/URDFParser.py
@@ -1,0 +1,246 @@
+import xml.etree.ElementTree as ET
+from collections import defaultdict
+from typing import Dict, List
+
+import numpy as np
+from scipy.spatial.transform import Rotation as ScipyRotation
+from trip_kinematics.Transformation import Transformation
+
+
+def from_urdf(filename: str) -> List[Transformation]:
+    """Generates a list of :py:class`Transformations` describing a robot specified in a URDF file.
+
+    Args:
+        filename (str): Path to URDF file.
+
+    Raises:
+        ValueError: Could not parse URDF file.
+
+    Returns:
+        List[Transformation]: List of transformations that describe the robot.
+    """
+    tree = ET.parse(filename)
+    root = tree.getroot()
+    joints = root.findall('joint')
+    joint_name_to_transformations = {}
+
+    joint_tree_dict = build_joint_tree_dict(joints)
+
+    root_joints = [name for name, value in joint_tree_dict.items() if not value['parent']]
+
+    for i, joint in enumerate(joints):
+        name = joint.get('name')
+        if not name:
+            raise ValueError('Missing name field in URDF file')
+        joint_tree_dict[name]['joint_index'] = i
+
+    for joint in joints:
+        joint_name_to_transformations[joint.get('name')] = get_transformations_for_joint(joint)
+
+    transformations = []
+    for joint in root_joints:
+        transformations.extend(create_transformations_from_tree(joint,
+                                                                joint_tree_dict,
+                                                                joint_name_to_transformations))
+
+    return transformations
+
+
+def build_joint_tree_dict(joints: List[ET.Element]) -> Dict[str, Dict]:
+    """Creates a dictionary representing parent-child relationships between joints (can be used to
+    build a tree of joints).
+
+    Args:
+        joints (List[ET.Element]): List of <joint> tags from the URDF file.
+
+    Returns:
+        dict[str, Dict]: A dictionary representing parent-child relationships between joints.
+    """
+    # Keep track of parent and children links of joints
+    # and use that to search for the parent joint of each joint
+    joint_tree_dict = {
+        joint.get('name'): {
+            'child_link': joint.find('child').get('link'),
+            'parent_link': joint.find('parent').get('link')
+        } for joint in joints}
+
+    parent_link_to_joint = defaultdict(list)
+    child_link_to_joint = defaultdict(list)
+
+    for joint in joints:
+        parent_link_to_joint[joint.find('parent').get('link')].append(joint.get('name'))
+        child_link_to_joint[joint.find('child').get('link')].append(joint.get('name'))
+
+    for val in joint_tree_dict.values():
+        val['parent'] = child_link_to_joint[val['parent_link']]
+        val['child'] = parent_link_to_joint[val['child_link']]
+
+    return joint_tree_dict
+
+
+def align_vectors(a: np.ndarray, b: np.ndarray) -> np.ndarray:
+    """Generates a rotation matrix that makes b parallel to a.
+
+    Args:
+        a (np.ndarray): 3D Vector.
+        b (np.ndarray): 3D Vector.
+
+    Returns:
+        np.ndarray: 3x3 rotation matrix.
+    """
+    a = a / np.linalg.norm(a)
+    b = b / np.linalg.norm(b)
+
+    # if the vectors are parallel but in opposite directions, return a 180 degree rotation
+    if np.array_equal(a, -b):
+        return -np.identity(3)
+
+    align_axis = np.cross(a, b)
+    cos_angle = np.dot(a, b)
+    k = 1 / (1 + cos_angle)
+
+    result = np.array([[(align_axis[0] * align_axis[0] * k) + cos_angle,
+                        (align_axis[1] * align_axis[0] * k) - align_axis[2],
+                        (align_axis[2] * align_axis[0] * k) + align_axis[1]],
[…]
