### 2021-12-13 6f95d82
Fix wrong assertions in urdf parser
