### 2021-12-13 2cddf47
Add environment.yml and notes
