+++ stuff/notes.md
@@ -1,0 +1,1 @@
+Should pytest be in setup.py (requirements)
