+++ stuff/environment.yml
@@ -1,0 +1,16 @@
+name: robotik
+
+dependencies:
+  - python=3.*
+  - pip
+  # casadi has a bug with newer versions of numpy
+  - numpy>=1.17.4, < 1.20.0
+  - autopep8
+  - pylint
+
+  - pip:
+    - casadi>=3.5.5
+    - TriP/src/
+    # forwards & inverse kinematics, for testing
+    - kinpy
+    - pytest
