### 2021-12-14 a4d4e93
Rephrase docstrings a bit
