### 2021-12-14 848965b
Add test for align_vectors()
