### 2021-12-23 0c6a770
Reraise exception explicitly
