### 2021-12-23 cd58916
Make class functions "private"
