### 2022-01-02 1840703
Clean up urdf parser test
