### 2022-01-05 6c10262
Fix bug from mutable default argument
