+++ tests/test_URDFParser.py
@@ -1,5 +1,6 @@
 import unittest
 import os
+import xml.etree.ElementTree as ET
 
 import kinpy as kp
 import numpy as np
@@ -19,8 +20,17 @@ def state_to_trip(state):
     }
 
 
+def printy(i, robot):
+    np.set_printoptions(precision=3, suppress=True)
+    print(robot.get_endeffectors()[i], '\n\n', trip_kinematics.forward_kinematics(
+        robot, robot.get_endeffectors()[i]).astype('float64'))
+
+
 class TestStates(unittest.TestCase):
     def test_simple_inverse_kinematics(self):
+        # TODO remove later
+        np.set_printoptions(precision=3, suppress=True)
+
         urdf_examples_dir = os.path.join('tests', 'urdf_examples')
         urdf_examples_filenames = os.listdir(urdf_examples_dir)
 
@@ -30,7 +40,8 @@ class TestStates(unittest.TestCase):
             with open(full_path, encoding='utf8') as file:
                 try:
                     # setup kinpy chain
-                    chain_kinpy = kp.build_chain_from_urdf(file.read())
+                    urdf_data_str = file.read()
+                    chain_kinpy = kp.build_chain_from_urdf(urdf_data_str)
                     state = {
                         joint_name: 0
                         for joint_name in chain_kinpy.get_joint_parameter_names()
@@ -41,21 +52,33 @@ class TestStates(unittest.TestCase):
                     chain_trip = trip_kinematics.URDFParser.from_urdf(full_path)
                     robot = trip_kinematics.Robot(chain_trip)
                     robot.set_actuated_state(state_to_trip(state))
-                    trip_kinematics.forward_kinematics(robot, 'elbow2_tra')
 
-                    forearm_kp = chain_kinpy.forward_kinematics(state_to_kinpy(state))['forearm']
-                    forearm_kp_rot_mat = trip_kinematics.Utility.quat_rotation_matrix(*forearm_kp.rot)
-                    forearm_kp_rot_hom = trip_kinematics.Utility.hom_rotation(forearm_kp_rot_mat)
-                    forearm_kp_tra_hom = trip_kinematics.Utility.hom_translation_matrix(*forearm_kp.pos)
+
+                    tree = ET.parse(full_path)
+                    root = tree.getroot()
+                    joints = root.findall('joint')
+                    joint_tree_dict = trip_kinematics.URDFParser._build_joint_tree_dict(joints)
+
+                    for joint in joint_tree_dict.values():
+                        transf_kp = chain_kinpy.forward_kinematics(state_to_kinpy(state)) \
+                            [joint['child_link']]
+
+                        transf_kp_hom = \
+                            trip_kinematics.Utility.hom_translation_matrix(*transf_kp.pos) \
+                            .astype('float64')
+
+                        transf_trip_hom = \
+                            trip_kinematics.forward_kinematics(robot, joint).astype('float64')
+
+                        assert np.allclose(transf_kp_hom, transf_trip_hom)
+
+                        print(joint['child_link'])
 
                     print(1)
 
                 except KeyError:
-                    # kinpy does not support planar or floating joints, don't run kinpy test
-                    # ? URDFToCasadi instead?
-                    pass
-
-        print('hi')
+                    print(
+                        f'Warning: planar and/or floating joint in {filename}; these are not currently supported. Skipping test case.')
 
     def test_align_vectors(self):
         test_cases = [
