### 2022-01-05 aeb7502
renamed the test to proper test name
