### 2022-01-05 22fbb84
changed loop to iterate over items, not values
