### 2022-01-09 #3967 @SeungBack
I reviewed the paper and documentation, and here are the comments. I am reviewing the codes and will add detailed comments about the codes soon.

This paper introduces an easy-to-use Python package for computing the forward and inverse kinematics of robots. The author claims that previous libraries lack support for hybrid serial-parallel systems, and TirP can be used to model both parallel and serial mechanisms. Solving Inverse and forward kinematics is an essential component for robotics, and I think this library can be useful for those who want to easily model the kinematics of complex robots.

Major issues
- There are other libraries that support computing the F.K and I.K of robots, such as [ikpy](https://github.com/Phylliade/ikpy) and [Klampt](https://github.com/krishauser/Klampt), but the paper compares TriP with Openrave, MoveIt, and Matlab. It would be better to describe the strength of this library by comparing it to other libraries in terms of functionality and speed.
- For the easy modeling of chain models, support of URDF import/export and 2D/3D visualization is necessary. Is there a plan to support this?

Minor issues
- I think the summary section does not summarize the paper, rather, it introduces the overall background.
- In the [tutorial](https://trip-kinematics.readthedocs.io/en/main/tutorials.html), some code references are broken and refer to the wrong lines of code
- Examples do not include the codes for importing packages. Please add this information (e.g. import trip_kinematics, from math import radians)
-Documentation includes only TriPed robots, and I think adding an example for popular serial mechanisms (e.g., Panda, UR5) would be helpful.
-The names of the pip package name (trip-kinematics) and paper title (TriP) do not match, and thus please consider adding an installation section on readthedocs.
- The colors of Figures 1 and 3 in the paper are confusing, and it would be better to use contrastive colors.
- typos in Figure 2. (e.g., ,,Spherical Joint", [,tz'])
