### 2022-01-10 998df5c
fixed wrong code lines for triped tutorial
