+++ docs/source/tutorials.rst
@@ -57,14 +57,14 @@ This leads to the following virtual open chain:
 .. literalinclude:: ../../src/trip_robots/triped_leg.py
    :language: python
    :linenos:
-   :lines: 153-158
+   :lines: 157-162
 
 And the corresponding transformations for the open chain:
 
 .. literalinclude:: ../../src/trip_robots/triped_leg.py
    :language: python
    :linenos:
-   :lines: 167-179
+   :lines: 171-183
 
 
 Note that both chains are made up of transformations without `state_variables`.
@@ -101,7 +101,7 @@ The final code can be seen down below:
 .. literalinclude:: ../../src/trip_robots/triped_leg.py
    :language: python
    :linenos:
-   :lines: 15-151
+   :lines: 15-155
 
 
 Building the group
@@ -112,7 +112,7 @@ Using both the mappings and the virtual open chain, the group can be build:
 .. literalinclude:: ../../src/trip_robots/triped_leg.py
    :language: python
    :linenos:
-   :lines: 160-165
+   :lines: 164-169
 
 Note that the closed chain specifies no parent since it is located directly at the robots base and the open chain specifies no mappings since these are autogenerated.
 
@@ -138,11 +138,11 @@ This function can be seen here:
 .. literalinclude:: ../../src/trip_robots/triped.py
    :language: python
    :linenos:
-   :lines: 10-79
+   :lines: 10-91
 
 The final TriPed robot can then be build using:
 
 .. literalinclude:: ../../src/trip_robots/triped.py
    :language: python
    :linenos:
-   :lines: 82
+   :lines: 94
