### 2022-01-10 #3967 @liquidcronos
Hi @SeungBack Thanks for your valuable feedback!
I am not sure If I should immediately respond or wait for the second reviewer, So I hope that it is alright to address the Issues:

> There are other libraries that support computing the F.K and I.K of robots, such as ikpy and Klampt, but the paper compares TriP with Openrave, MoveIt, and Matlab. It would be better to describe the strength of this library by comparing it to other libraries in terms of functionality and speed.

You are correct we do not mention other python kinematic libraries.
This was an oversight, ultimately the libraries you mentioned suffer from the same shortcomings as the other ones namely lacking support for serial mechanisms or hybrid serial parallel mechanisms.
I will add them shortly.

> For the easy modeling of chain models, support of URDF import/export and 2D/3D visualization is necessary. Is there a plan to support this?

We completely agree that  visualization is very important in practice.
For this reason we are working on adding visualizations for a future release of the library.

While URDF is a standard format for robot descriptions, it is unfortunately rather basic and lacks support for the hybrid mechanisms TriP aims to model.
Since it is still sufficient to describe the virtual open chain of a mechanism we are currently working on URDF support which we will hopefully finish next month.

> I think the summary section does not summarize the paper, rather, it introduces the overall background.

You are right. To be honest I was a bit confused by the required content of this section since many other papers seem to include the background in order to describe the purpose of the software for a 'diverse, non-specialist audience'.
Would you rather I shorten the overview or write more about TriP itself?

> In the tutorial, some code references are broken and refer to the wrong lines of code

> Examples do not include the codes for importing packages. Please add this information (e.g. import trip_kinematics, from math import radians)

Thanks for letting me know, should be fixed now :)

> Documentation includes only TriPed robots, and I think adding an example for popular serial mechanisms (e.g., Panda, UR5) would be helpful.

Ideally we want more examples of hybrid serial parallel robots to better showcase TriPs performance as you can see [here](https://github.com/TriPed-Robot/trip_kinematics/issues/57).
Popular examples of such hybrid systems are the Kuka KR 40 PA or the fanuc-m410ic.
The main reason we have not included these systems so far is that we want to independently verify that we have correctly implemented the models.
Since we are the only library supporting hybrid mechanisms calculating reference values is very difficult.
For this reason we have currently included the excavator robots as toy examples while we slowly add more examples.
If you want we can prioritize the development of these robot models to get them out sooner.

> The names of the pip package name (trip-kinematics) and paper title (TriP) do not match, and thus please consider adding an installation section on readthedocs.

Good point, we will add a getting started section including the installation instructions of the README.md.

> The colors of Figures 1 and 3 in the paper are confusing, and it would be better to use contrastive colors.

Are you referring only to the two green colors, or the blue and green color as well?
I will try to change them to be more contrastive.
