### 2022-01-10 7238791
adding default values to the optional urdf tags
