### 2022-01-10 f6bd6d0
comented out floating and planar as these are still unsupported
