### 2022-01-10 75234a6
creating the state now based on trip actuated state instead of kinpy
