### 2022-01-10 b817efa
fixing the test to consider kinpy rotation and not just translation
