### 2022-01-10 96e5421
exported functionality to individual functions from test_all_urdf_files
