### 2022-01-10 bdb83f0
added comments for future urdf tests and added test for large tree
