### 2022-01-10 4addf61
Fix linter complaints & formatting
