+++ paper.bib
@@ -1,5 +1,21 @@
 % Encoding: UTF-8
 
+@online{Klampt,
+  title={Kris' Locomotion and Manipulation Planning Toolbox - Klamp't},
+  url={https://github.com/krishauser/Klampt},
+  urldate = {2021-03-31},
+  author ={Kris Hauser}
+}
+
+
+@online{ikpy,
+  title={IKPy An Inverse Kinematics library aiming performance and modularity},
+  url={https://github.com/Phylliade/ikpy},
+  urldate = {2021-07-10},
+  author ={Pierre Manceron}
+}
+
+
 
 
 @Article{survey,
