+++ paper.bib
@@ -4,6 +4,7 @@
   title={Kris' Locomotion and Manipulation Planning Toolbox - Klamp't},
   url={https://github.com/krishauser/Klampt},
   urldate = {2021-03-31},
+  year    = {2021},
   author ={Kris Hauser}
 }
 
@@ -12,6 +13,7 @@
   title={IKPy An Inverse Kinematics library aiming performance and modularity},
   url={https://github.com/Phylliade/ikpy},
   urldate = {2021-07-10},
+  year    = {2021},
   author ={Pierre Manceron}
 }
 
