### 2022-01-12 fae7dd0
Add simple visualizer
