### 2022-01-13 #3967 @SeungBack
Hi @liquidcronos,

I have checked the updated version of the manuscripts, and I am satisfied with the response and the revised manuscript. It is good to hear that TriP will be updated to support URDF and visualization. Here are the additional comments for your response:

> 1. Would you rather I shorten the overview or write more about TriP itself?

I recommend adding more about TriP itself (e.g., features, structures, and the strength of TriP).

> 2. Are you referring only to the two green colors, or the blue and green colors as well?

The colors of green and blue links are hard to distinguish for me.

> 3. Examples do not include the codes for importing packages. Please add this information (e.g. import trip_kinematics, from math import radians)
>> Thanks for letting me know, should be fixed now :)

Please let me know after updating the codes so that I can run and review them.
