### 2022-01-13 2e019de
made urdf_examples_dir a global variable and run_test returns T/F now
