### 2022-01-13 1b62b71
switched the rot from euler angles to quaternions to prevent rot errors
