### 2022-01-13 23777d8
Move test cases into one function
