### 2022-01-13 77b68ac
Do a bit of refactoring + pep8 stuff
