+++ tests/test_URDFParser.py
@@ -4,34 +4,29 @@ import xml.etree.ElementTree as ET
 
 import kinpy as kp
 import numpy as np
-# from trip_kinematics.URDFParser import align_vectors
+from trip_kinematics import Utility
 import trip_kinematics.URDFParser
 import trip_kinematics.Robot
 
 urdf_examples_dir = os.path.join('tests', 'urdf_examples')
 
 
-def state_to_kinpy(state):
+def to_kp(state):
     return {
         "_".join(joint_name.split("_")[:-2]): value
         for joint_name, value in state.items()
     }
 
 
-def state_to_trip(state):
+def to_trip(state):
     return state
 
 
-def get_state(robot):
+def initialize_state(robot):
     return {
         joint_name: 0
         for joint_name in robot.get_actuated_state()
     }
-
-
-def create_trip_robot(path):
-    chain_trip = trip_kinematics.URDFParser.from_urdf(path)
-    return trip_kinematics.Robot(chain_trip)
 
 
 def create_kinpy_chain(path):
@@ -49,46 +44,39 @@ def get_joint_tree_dict(path):
     return trip_kinematics.URDFParser._build_joint_tree_dict(joints)
 
 
-def run_test(path):
-    # setup TriP robot using the URDF parser
-    # this needs a try catch
+def compare_urdf_trip_vs_kinpy(path):
     path = os.path.join(urdf_examples_dir, path)
-    robot = create_trip_robot(path)
-    state = get_state(robot)
 
-    # setup kinpy chain
-    # this needs a try catch
-    chain_kinpy = create_kinpy_chain(path)
-    chain_kinpy.forward_kinematics(state_to_kinpy(state))
+    # Setup TriP robot using the URDF parser
+    try:
+        robot = trip_kinematics.Robot(trip_kinematics.URDFParser.from_urdf(path))
+        state_init = initialize_state(robot)
+        robot.set_actuated_state(to_trip(state_init))
+    except ValueError as err:
+        raise ValueError(f'File {path} contains unupported joint type ({err})') from err
 
-    # this needs a try catch?
-    robot.set_actuated_state(state_to_trip(state))
+    # Setup kinpy chain
+    try:
+        chain_kinpy = create_kinpy_chain(path)
+        chain_kinpy.forward_kinematics(to_kp(state_init))
+    except KeyError as err:
+        raise ValueError(f'File {path} contains unupported joint type ({err})') from err
 
-    joint_tree_dict = get_joint_tree_dict(path)
+    for joint, joint_dict in get_joint_tree_dict(path).items():
+        # Calculate homogenous transformation matrix of kinpy forward kinematics
+        transf_kp = chain_kinpy.forward_kinematics(to_kp(state_init))[joint_dict['child_link']]
 
-    for joint, joint_dict in joint_tree_dict.items():
-        # calculate homogenous matrix of kinpy forward kinematic for every joint
-        transf_kp = chain_kinpy.forward_kinematics(state_to_kinpy(state))[joint_dict['child_link']]
-
-        transf_kp_hom_pos = \
-            trip_kinematics.Utility.hom_translation_matrix(*transf_kp.pos) \
-            .astype('float64')
-        transf_kp_hom_rot = \
-            trip_kinematics.Utility.hom_rotation(
-                trip_kinematics.Utility.quat_rotation_matrix(*transf_kp.rot).astype('float64')
-            ) \
-            .astype('float64')
+        kp_rot_matrix = Utility.quat_rotation_matrix(*transf_kp.rot).astype('float64')
+        transf_kp_hom_rot = Utility.hom_rotation(kp_rot_matrix).astype('float64')
+        transf_kp_hom_pos = Utility.hom_translation_matrix(*transf_kp.pos).astype('float64')
 
[…]
