### 2022-01-13 be1ca9d
Add pseudorandom test cases with fixed seed
