+++ tests/test_URDFParser.py
@@ -1,14 +1,12 @@
 import unittest
 import os
 import xml.etree.ElementTree as ET
+import random
 
 import kinpy as kp
 import numpy as np
-from trip_kinematics import Utility
-import trip_kinematics.URDFParser
-import trip_kinematics.Robot
-
-urdf_examples_dir = os.path.join('tests', 'urdf_examples')
+from trip_kinematics import Utility, URDFParser, Robot
+import trip_kinematics
 
 
 def to_kp(state):
@@ -41,72 +39,72 @@ def get_joint_tree_dict(path):
     joints = root.findall('joint')
 
     # pylint: disable=protected-access
-    return trip_kinematics.URDFParser._build_joint_tree_dict(joints)
+    return URDFParser._build_joint_tree_dict(joints)
 
 
-def compare_urdf_trip_vs_kinpy(path):
-    path = os.path.join(urdf_examples_dir, path)
-
+def compare_urdf_trip_vs_kinpy(path, rng_states_count=10):
     # Setup TriP robot using the URDF parser
     try:
-        robot = trip_kinematics.Robot(trip_kinematics.URDFParser.from_urdf(path))
+        robot = Robot(URDFParser.from_urdf(path))
         state_init = initialize_state(robot)
-        robot.set_actuated_state(to_trip(state_init))
     except ValueError as err:
-        raise ValueError(f'File {path} contains unupported joint type ({err})') from err
+        raise ValueError(f'File {path} contains unsupported joint type ({err})') from err
 
     # Setup kinpy chain
     try:
         chain_kinpy = create_kinpy_chain(path)
-        chain_kinpy.forward_kinematics(to_kp(state_init))
     except KeyError as err:
-        raise ValueError(f'File {path} contains unupported joint type ({err})') from err
+        raise ValueError(f'File {path} contains unsupported joint type ({err})') from err
 
-    for joint, joint_dict in get_joint_tree_dict(path).items():
-        # Calculate homogenous transformation matrix of kinpy forward kinematics
-        transf_kp = chain_kinpy.forward_kinematics(to_kp(state_init))[joint_dict['child_link']]
+    test_states = [state_init]
 
-        kp_rot_matrix = Utility.quat_rotation_matrix(*transf_kp.rot).astype('float64')
-        transf_kp_hom_rot = Utility.hom_rotation(kp_rot_matrix).astype('float64')
-        transf_kp_hom_pos = Utility.hom_translation_matrix(*transf_kp.pos).astype('float64')
+    for seed in range(rng_states_count):
+        random.seed(seed)
+        new_state = {
+            joint: random.uniform(-np.pi, np.pi)
+            for joint in state_init.keys()
+        }
+        test_states.append(new_state)
 
-        transf_kp_hom = transf_kp_hom_pos @ transf_kp_hom_rot
+    for state in test_states:
+        robot.set_actuated_state(to_trip(state))
+        chain_kinpy.forward_kinematics(to_kp(state))
 
-        # Calculate homogenous matrix of TriP forward kinematic for every joint
-        transf_trip_hom = trip_kinematics.forward_kinematics(robot, joint).astype('float64')
+        # Calculate homogenous transformation matrices using both kinpy and TriP for every joint
+        # and compare them. joint_dict contains info about connections to other joints.
+        for joint, joint_dict in get_joint_tree_dict(path).items():
+            # kinpy
+            transf_kp = chain_kinpy.forward_kinematics(to_kp(state))[joint_dict['child_link']]
+            kp_rot_matrix = Utility.quat_rotation_matrix(*transf_kp.rot).astype('float64')
+            transf_kp_hom_rot = Utility.hom_rotation(kp_rot_matrix).astype('float64')
[…]
