### 2022-01-13 214ba58
changed the default axis to 0 0 1 to acomodate to kinpy
