### 2022-01-13 830186d
removed duplicate assert statement
