### 2022-01-13 82d896a
added error related tests
