### 2022-01-13 b66b573
Improve pep8 compatibility + refactor
