+++ stuff/notes.md
@@ -1,1 +1,0 @@
-Should pytest be in setup.py (requirements)
