### 2022-01-13 4c2812e
Add kinpy and defusedxml to setup.py dependencies
