### 2022-01-14 58d2745
Add more comments, docs, clean up
