+++ src/trip_kinematics/URDFParser.py
@@ -10,6 +10,9 @@ from trip_kinematics.Transformation import Transformation
 def from_urdf(filename: str) -> List[Transformation]:
     """Converts a robot specified in a URDF file into a list of :py:class`Transformation` objects.
 
+    If the <origin> tag does not specify xyz and rpy values or the tag is omitted, these default
+    to (0, 0, 0). <axis> defaults to (0, 0, 1).
+
     Args:
         filename (str): Path to URDF file.
 
@@ -115,10 +118,12 @@ def _build_joint_tree_dict(joints: List[ET.Element]) -> Dict[str, Dict]:
 
 def _get_transformations_for_joint(joint: ET.Element) -> List[List]:
     """Generates the parameters for the transformations for the input joint. One joint is
-    represented by up to four transformations. These are: 1. translation and 2. rotation (both taken
-    from the <origin> tag of the URDF file), then 3. a rotation which ensures joint movement is
-    aligned with the axes of the local coordinate system, then 4. a transformation which represents
-    joint movement (the only non-fixed transformation of the four).
+    represented by up to five transformations. These are:
+
+        1. origin translation and 2. origin rotation (both taken from the <origin> tag of the URDF),
+        3. a rotation that makes joint movement parallel to the z axis of the local coord. system,
+        4. the joint movement (the only non-fixed transformation of the five),
+        5. the inverse transformation of number 3.
 
     Args:
         joint (ET.Element): A <joint> tag in the URDF file.
@@ -136,11 +141,13 @@ def _get_transformations_for_joint(joint: ET.Element) -> List[List]:
     origin = joint.find('origin')
     joint_transformations = []
 
+    # File needs to specify a joint type for each joint
     try:
         assert type_ is not None
     except AssertionError as err:
         raise ValueError(f'Error: Invalid URDF file ({err})') from err
 
+    # Default values if origin rotation or translation are not specified
     if origin is None:
         xyz_vals = '0 0 0'
         rpy_vals = '0 0 0'
@@ -151,26 +158,30 @@ def _get_transformations_for_joint(joint: ET.Element) -> List[List]:
             xyz_vals = '0 0 0'
         if rpy_vals is None:
             rpy_vals = '0 0 0'
+    origin_xyz = np.array(list(map(float, xyz_vals.split(' '))))
+    origin_rpy = np.array(list(map(float, rpy_vals.split(' '))))
 
-    # Translation and rotation
-    xyz = np.array(list(map(float, xyz_vals.split(' '))))
-    rpy = np.array(list(map(float, rpy_vals.split(' '))))
+    if type_ in ['fixed']:
+        axis = None     # Fixed joints have no axis
 
-    # Fixed and floating joints have no axis
-    if type_ in ['fixed', 'floating']:
-        axis = None
-    else:
+    elif type_ in ['continuous', 'revolute', 'prismatic']:
         axis = joint.find('axis')
+        # Default to (0, 0, 1) if the axis is unspecified. Note that there does not seem to be
+        # a standard default value for the axis. Maybe better to remove the default to prevent
+        # potential confusion?
         if axis is None:
             axis = '0 0 1'
         else:
             axis = axis.get('xyz')
         axis = np.array(list(map(float, axis.split(' '))))
-        axis = axis / np.linalg.norm(axis)
+        axis /= np.linalg.norm(axis)
 
-    # For each joint, define four transformations
-    tra = [name + '_tra', {'tx': xyz[0], 'ty': xyz[1], 'tz': xyz[2]}, []]
-    rot_quat = ScipyRotation.from_euler('xyz', [rpy[0], rpy[1], rpy[2]], degrees=False).as_quat()
[…]
