### 2022-01-18 055e896
expanded description of trips features
