### 2022-01-18 e243451
added excavator trafos in second example
this should keep the closure equation example self sufficient
