### 2022-01-18 f5181c5
Added docstrings to functions in test_URDFParser.py
