### 2022-01-18 66c0537
updated align_vectors docstring
