### 2022-01-18 d5fffe8
updated _get_transformations_for_joint docstring non-fixed -> dynamic
