### 2022-01-18 41de1df
moved the assertion of joint type up towards the top of the function
