### 2022-01-18 08c37a0
replaced if statement for the mov definition with a dictionary call
