### 2022-01-19 cdc8130
added getting started section to TriPs documentation
