+++ docs/source/getting_started.rst
@@ -1,0 +1,23 @@
+###############
+Getting Started
+###############
+
+The current release of the project can always be found on the python package index `PyPi <https://en.wikipedia.org/wiki/Quaternion>`_ .
+However, since another python package is already called trip, the package is called trip_kinematics.
+The current stable release can be installed using:
+
+::
+
+    pip install trip_kinematics
+
+
+Alternatively, the current development version can be downloaded using Github:
+
+::
+
+    git clone https://github.com/TriPed-Robot/TriP
+    cd TriP
+    pip install src/
+
+
+For more information on how to use the library once installed visit the Tutorials or read 'How TriP models Robots'.
