### 2022-01-19 #3967 @liquidcronos
Hey @SeungBack
I have amended the paper with more references to TriPs features.
As promised I have also added a getting started section that explains how the library is installed.
The code examples should now also be self-contained and include all necessary imports.

Are there any other changes you would like to see apart from the figure colors (which I should be able to change until Friday)?
