### 2022-01-21 76ca584
Change some docstrings and formatting stuff
