+++ tests/test_URDFParser.py
@@ -1,6 +1,6 @@
 import unittest
 import os
-import xml.etree.ElementTree as ET
+import defusedxml.ElementTree as ET
 import random
 
 import kinpy as kp
