+++ src/setup.py
@@ -7,6 +7,6 @@ setuptools.setup(
     author='Torben Miller, Jan Baumgärtner',
     license='MIT',
     description='...',
-    install_requires=['casadi>=3.5.5', 'numpy>=1.17.4, < 1.20.0', 'kinpy'],
+    install_requires=['casadi>=3.5.5', 'numpy>=1.17.4, < 1.20.0', 'defusedxml>=0.5'],
     packages=['trip_kinematics', 'trip_robots']
 )
