### 2022-01-21 8af1e4b
Add better docstrings
