### 2022-01-21 5090f9c
Add test for errors
