### 2022-01-21 c1d15c8
Add function for saving forward kin to json
