### 2022-01-21 17b0fae
started implementing tests that pull from json file to compare frwrd kin
