+++ tests/test_URDFParser.py
@@ -1,12 +1,12 @@
 import unittest
 import os
-import defusedxml.ElementTree as ET
-import random
+import pathlib
 import json
 
+import defusedxml.ElementTree as ET
 import kinpy as kp
 import numpy as np
-from trip_kinematics import Utility, URDFParser, Robot, forward_kinematics as trip_forward_kin
+from trip_kinematics import Utility, URDFParser, Robot, forward_kinematics
 
 
 def state_to_kp(state):
@@ -23,10 +23,6 @@ def state_to_kp(state):
 
     """
     return state
-    # return {
-    #     "_".join(joint_name.split("_")[:-2]): value
-    #     for joint_name, value in state.items()
-    # }
 
 
 def state_to_trip(trip_state, kp_state):
@@ -92,6 +88,7 @@ def get_joint_tree_dict(path):
 
     Returns:
         Dict[str, Dict]: Dictionary representing parent-child relationships between joints.
+
     """
     tree = ET.parse(path)
     root = tree.getroot()
@@ -101,17 +98,15 @@ def get_joint_tree_dict(path):
     return URDFParser._build_joint_tree_dict(joints)
 
 
-def compare_urdf_trip_vs_kinpy(path, rng_states_count=10, atol=1e-08):
-    """Reads a URDF file, converts it into a TriP robot and a kinpy kinematic chain, calculates
-    forward kinematics for both, and checks whether the results are within a tolerance of each
-    other. This is done first using the state where all joint positions are set to zero, then for
-    a number of states where all joint positions are generated randomly.
+def compare_urdf_trip_vs_json(path, atol=1e-08):
+    """Reads a URDF file, converts it into a TriP robot, calculates forward kinematics, and checks
+    whether the results are close to those saved in a JSON file. The JSON file must have the
+    same name as the URDF file, except the extension should be changed from .urdf to .json.
 
     Args:
         path (str): Path to the URDF file to test.
-        rng_states_count (int, optional): The number of randomized states. Defaults to 10.
-        atol (float, optional): The absolute tolerance parameter for numpy :py:func:`allclose`.
-                       Defaults to 1e-08
+        atol (float, optional): Absolute tolerance of the comparison; see numpy documentation for
+                                :py:func:`allclose`. Defaults to 1e-08.
 
     Raises:
         AssertionError: Results were not within tolerance of each other.
@@ -126,76 +121,28 @@ def compare_urdf_trip_vs_kinpy(path, rng_states_count=10, atol=1e-08):
             f'File {path} is not valid. Unsupported joint type? (error was {err})'
         ) from err
 
-    # Setup kinpy chain
-    try:
-        chain_kinpy = create_kinpy_chain(path)
-    except KeyError as err:
-        raise ValueError(
-            f'File {path} is not valid. Unsupported joint type? Missing tag? (error was {err})'
-        ) from err
-
-    path_json = path[:-4] + 'json'
-    with open(path_json) as json_file:
-        data = json.load(json_file)
-        # states = data[0]['state']
-        # transformations = data[0]['transformations']
-        # print(states)
-        # print("------")
-        # print(transformations)
-
-        # go through every generated set of states and compare them to what trip calculates
-        for i, robot_position in enumerate(data):
-            # first set the state from the file to TriP
-            current_state = robot_position['state']
-            current_transformations = robot_position['transformations']
-            robot.set_actuated_state(state_to_trip(state_init, current_state))
+    path_json = pathlib.Path(path).with_suffix('.json')
+    with open(path_json, encoding='utf8') as json_file:
[…]
