### 2022-01-22 f28cef1
Delete useless functions
