### 2022-01-22 3032375
Delete useless functions
