### 2022-01-23 7279eae
removed function call that was left in as a test
