### 2022-01-23 5532643
added dummy floating and planar joints, they act as fixed
