### 2022-01-23 1f4ddb5
Rename joint_(tree_)dict to joint_info
