### 2022-01-23 5c4a4a8
Move precomputed kinematics to their own folder
