### 2022-01-23 e2005b9
Move urdf_examples_dir to a variable
