### 2022-01-26 efc9cb0
Add Rotation class
