### 2022-01-27 ed39808
Add Rotation class and fix issues
