+++ tests/test_utility.py
@@ -13,7 +13,8 @@ class TestStates(unittest.TestCase):
 
     def test_from_euler_to_quat(self):
         from_euler_cases = [
-            ([1, 2, 3], [0.4359528440735657, -0.7182870182434115, 0.3106224510657039, 0.44443511344300074]),
+            ([1, 2, 3],
+             [0.4359528440735657, -0.7182870182434115, 0.3106224510657039, 0.44443511344300074]),
             ([1, 0, 0], [0.8775825618903728, 0.479425538604203, 0.0, 0.0]),
             ([0, 1, 0], [0.8775825618903728, 0.0, 0.479425538604203, 0.0]),
             ([0, 0, 1], [0.8775825618903728, 0.0, 0.0, 0.479425538604203]),
