### 2022-02-01 d4f734a
Remove scipy import
