### 2022-02-04 #3967 @SeungBack
Hi @liquidcronos,

> Examples do not include the codes for importing packages. Please add this information (e.g. import trip_kinematics, from math import radians)

I still can not find the codes for importing other libraries in [Tutorial](https://trip-kinematics.readthedocs.io/en/main/tutorials.html#building-a-robot-model) section. Where can I find the information about it?
