### 2022-02-04 69d512f
added imports to triped tutorial
