### 2022-02-07 #3967 @danielskatz
Hi - I'm just checking in on this review.

@bmagyar - it doesn't look like you've gotten started on your checklist.  How are things going?

@SeungBack - thanks for your progress so far.

@liquidcronos - thanks for responding to issues/comments

Let me know if there are any problems I can help with.
