### 2022-02-18 TriPed-Robot/trip_kinematics#72 @bmagyar
Typo in figure
`virtual_state = {revoluze:`

![Screenshot from 2022-02-18 06-56-43](https://user-images.githubusercontent.com/3524577/154633308-1339fbe9-e393-41e9-bd91-ddd97797e353.png)
