+++ README.md
@@ -30,7 +30,7 @@ Installing the latest release (recommended):
 pip install trip-kinematics
 ```
 
-Installing the lnewest version from source:
+Installing the newest version from source:
 ```
 git clone https://github.com/TriPed-Robot/TriP
 cd TriP
