### 2022-03-01 TriPed-Robot/trip_kinematics#74 @bmagyar
Tutorials don't seem to work on Ubuntu 20, Python 3.8
At least I'm trying a blunt copy paste and getting a bunch of issues from casadi and sometimes just missing bits.

```
NameError                                 Traceback (most recent call last)
<ipython-input-10-b74461665e3c> in <module>
      3                         values={'tx': 1.640, 'tz': -0.037, },
      4
----> 5                         parent=closed_chain)
      6
      7 a_ll_zero = Transformation(name='zero_angle_convention',

NameError: name 'closed_chain' is not defined
```
and
```
Exception: Implicit conversion of symbolic CasADi type to numeric matrix not supported.
This may occur when you pass a CasADi object to a numpy function.
Use an equivalent CasADi function instead of that numpy function.
```

It may be possible it is not expected to work on the above setup.

*Please specify* in the tutorials section what setup the tutorials were tested on: OS & Python version should be enough.
