### 2022-03-01 #3967 @bmagyar
Great work guys!

I really only have minor things related to docs and the examples:
https://github.com/TriPed-Robot/trip_kinematics/issues/72 <-- this has been pending for 12 days or so
https://github.com/TriPed-Robot/trip_kinematics/issues/74 <-- this one being the main one
https://github.com/TriPed-Robot/trip_kinematics/pull/73 <-- miniscule fix submitted by me ;)
