### 2022-03-01 TriPed-Robot/trip_kinematics#74 @danielskatz
https://github.com/openjournals/joss-reviews/issues/3967 (tracking)
