### 2022-03-02 #3967 @liquidcronos
Thanks for your review @bmagyar !
I seem to have set my notification settings wrong, because I wasn't informed of your Issues by Mail.
This should be rectified now.

I will fix the type you found in  https://github.com/TriPed-Robot/trip_kinematics/issues/72 right away.
Thanks also for spotting the typo in the install instructions, I have approved your pull request and even found another typo in the same section...

Regarding your Issues with the tutorials, I have addressed them in the issue.
Please let me know if these fixed are sufficient or if you still want more detailed instructions on the required version of the tutorial.
