### 2022-03-15 TriPed-Robot/trip_kinematics#74 @bmagyar
Thanks for the response!

Snippet: looks clear enough now, thanks! I do recommend taking a second look at the url as it seems markdown didn't render it as one...

Casadi error: indeed the numpy version that produced it on my machine was 1.22.2, thanks for the correction.
