+++ paper.md
@@ -29,42 +29,41 @@ Serial mechanisms like robotic arms are mechanisms where each moving part (calle
 They are often used when a large workspace is required, meaning the robot needs a long reach.
 In parallel mechanisms, the links of the robot form loops causing them to be structurally stronger and stiffer.
 
-If both a large workspace and structural strength are required, hybrids between serial and parallel mechanisms are used.
-These mechanisms contain both serial and parallel mechanisms.
-While hybrid mechanisms combine the mechanical advantages of both parallel and serial mechanisms they also combine their modeling disadvantages:
+If both a large workspace and structural strength are required, hybrids that contain both serial and parallel mechanisms are used.
+While hybrid mechanisms combine the mechanical advantages of both parallel and serial mechanisms, they also combine their modeling disadvantages:
 
-​Finding an explicit solution for either forward or inverse kinematics is often impossible.
-Using numerical approaches instead leads to complicated constrained optimization problems for both forward and inverse kinematics. ​
+* Finding an explicit solution for either forward or inverse kinematics is often impossible.
+Using numerical approaches instead leads to complicated constrained optimization problems for both forward and inverse kinematics.
 
-While serial mechanisms are very well supported by current robotic frameworks, parallel mechanisms and hybrid mechanisms especially are often not supported at all.
+* While serial mechanisms are very well supported by current robotic frameworks, parallel mechanisms and hybrid mechanisms especially are often not supported at all.
 A great overview of the supported robot types for different robotic frameworks was compiled by @kumar.
 
-​TriP is a python package designed to close this gap using a modular modeling framework akin to the one described by ​@survey​.
-​It allows the modeling of arbitrary hybrid mechanisms and is capable of calculating forward and inverse kinematics.
+TriP is a python package designed to close this gap using a modular modeling framework akin to the one described by @survey.
+It allows the modeling of arbitrary hybrid mechanisms and is capable of calculating forward and inverse kinematics.
 
 The calculations are performed using a symbolic framework.
 This makes it easy for users to implement custom case-dependent mathematical solvers.
 
 # Statement of Need
-While a huge number of researchers use hybrid serial parallel systems such as @PISLA, @verne, or @berkley, most modern kinematics frameworks still lack support.
-Examples include openrave [@openrave] used in the moveit stack [@moveit], the matlab robotics toolbox [@matlabrobot], Klampt [@Klampt] or the inverse kinematics python library [@ikpy].
-While all these frameworks and libraries offer fast computation of forward and inverse kinematics they only support branching serial mechanisms.
-​This lack of support often leaves developers with essentially two choices:
+While a huge number of researchers, such as @PISLA, @verne, and @berkley, use hybrid serial parallel systems, most modern kinematics frameworks still lack support for such systems.
+Examples include OpenRAVE [@openrave], used in the MoveIt! stack [@moveit], the Matlab robotics toolbox [@matlabrobot], Klampt [@Klampt], and the inverse kinematics Python library [@ikpy].
+While all these frameworks and libraries offer fast computation of forward and inverse kinematics, they only support branching serial mechanisms.
+This lack of support often leaves developers with essentially two choices:
 
 Either shoehorn their hybrid robots into a framework not designed to handle them or be left to implement their own kinematic solvers.
 
-​TriP is a lightweight and easy-to-use package directly modeling hybrid mechanisms and calculating their kinematics.
-Although TriP is fast for a python package it is not built for robust hard-real time control applications.
+TriP is a lightweight and easy-to-use package for directly modeling hybrid mechanisms and calculating their kinematics.
+Although TriP is fast for a Python package, it is not built for robust hard-real time control applications.
[…]
