+++ paper.bib
@@ -1,7 +1,7 @@
 % Encoding: UTF-8
 
 @online{Klampt,
-  title={Kris' Locomotion and Manipulation Planning Toolbox - Klamp't},
+  title={Kris' Locomotion and Manipulation Planning Toolbox - {K}lamp't},
   url={https://github.com/krishauser/Klampt},
   urldate = {2021-03-31},
   year    = {2021},
@@ -10,7 +10,7 @@
 
 
 @online{ikpy,
-  title={IKPy An Inverse Kinematics library aiming performance and modularity},
+  title={IKPy - An Inverse Kinematics library aiming performance and modularity},
   url={https://github.com/Phylliade/ikpy},
   urldate = {2021-07-10},
   year    = {2021},
@@ -44,9 +44,9 @@
 
 
 @Article{berkley,
-  author  = {Zoss, A.B. and Kazerooni, H. and Chu, A.},
-  journal = {Mechatronics, IEEE/ASME Transactions on},
-  title   = {Biomechanical Design of the Berkeley Lower Extremity Exoskeleton (BLEEX)},
+  author  = {Zoss, A. B. and Kazerooni, H. and Chu, A.},
+  journal = {IEEE/ASME Transactions on Mechatronics},
+  title   = {Biomechanical Design of the {B}erkeley Lower Extremity Exoskeleton (BLEEX)},
   year    = {2006},
   month   = {05},
   pages   = {128 - 138},
@@ -123,7 +123,7 @@ note="www.coppeliarobotics.com"
 }
 
 @article{moveit,
-  title={Moveit![ros topics]},
+  title={Moveit! [{ROS} topics]},
   author={Chitta, Sachin and Sucan, Ioan and Cousins, Steve},
   journal={IEEE Robotics \& Automation Magazine},
   volume={19},
