### 2022-03-15 #3967 @danielskatz
👋 @liquidcronos - I've suggested a bunch of changes in https://github.com/TriPed-Robot/trip_kinematics/pull/75  If these are ok, please merge them. If not, we can discuss.  Then we can move to final acceptance and publication.
