### 2022-03-16 #3967 @danielskatz
@liquidcronos - can you please change the metadata of the zenodo archive to have "TriP: A Python package for the kinematic modeling of serial-parallel hybrid robots" as the title?  (See "How can I edit the metadata of a published record?" on https://help.zenodo.org)
