### 2022-03-16 #3967 @liquidcronos
@danielskatz Should be updated know!
By the way thanks for making me aware of the edit button, Up until know I was changing the .zenodo.json file by hand and drafting a new release each time...
