### 2022-03-21 #3967 @liquidcronos
Hey Guys, sorry for the late response, I was on vacation.
I just wanted to quickly thank everyone involved with this paper, although the process took longer than expected I really enjoyed how open and transparent this process was.
Special thanks also to you @danielskatz for being the editor even though you are normally not an editor in the robotics domain.
